fix(install): Add checks for existing oh-my-zsh, tmux, and vim dirs
With this change, the install script should now be capable of updating they system.

+++ install.sh
@@ -60,26 +60,34 @@ sudo -u $SUDO_USER pip3 install thefuck
 sudo -u $SUDO_USER pip3 install commitizen
 
 # oh-my-zsh setup 
-sudo -u $SUDO_USER sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" "" --unattended
-sudo -u $SUDO_USER git clone https://github.com/zsh-users/zsh-autosuggestions /home/$SUDO_USER/.oh-my-zsh/custom/plugins/zsh-autosuggestions
-sudo -u $SUDO_USER git clone https://github.com/zsh-users/zsh-history-substring-search /home/$SUDO_USER/.oh-my-zsh/custom/plugins/zsh-history-substring-search
-sudo -u $SUDO_USER git clone https://github.com/zsh-users/zsh-syntax-highlighting.git /home/$SUDO_USER/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting
-sudo -u $SUDO_USER rm /home/$SUDO_USER/.zshrc
-
+if [ ! -d "/home/$SUDO_USER/.oh-my-zsh" ]
+then 
+    sudo -u $SUDO_USER sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" "" --unattended
+    sudo -u $SUDO_USER git clone https://github.com/zsh-users/zsh-autosuggestions /home/$SUDO_USER/.oh-my-zsh/custom/plugins/zsh-autosuggestions
+    sudo -u $SUDO_USER git clone https://github.com/zsh-users/zsh-history-substring-search /home/$SUDO_USER/.oh-my-zsh/custom/plugins/zsh-history-substring-search
+    sudo -u $SUDO_USER git clone https://github.com/zsh-users/zsh-syntax-highlighting.git /home/$SUDO_USER/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting
+    sudo -u $SUDO_USER rm /home/$SUDO_USER/.zshrc
+fi
 # Vim setup
-sudo -u $SUDO_USER git clone https://github.com/VundleVim/Vundle.vim.git /home/$SUDO_USER/.vim/bundle/Vundle.vim
-sudo -u $SUDO_USER vim +visual +PluginInstall +qall
+if [ ! -d "/home/$SUDO_USER/.vim"]
+then
+    sudo -u $SUDO_USER git clone https://github.com/VundleVim/Vundle.vim.git /home/$SUDO_USER/.vim/bundle/Vundle.vim
+    sudo -u $SUDO_USER vim +visual +PluginInstall +qall
+fi
 
 # Tmux setup
-sudo -u $SUDO_USER git clone https://github.com/tmux-plugins/tpm /home/$SUDO_USER/.tmux/plugins/tpm
-# start a server but don't attach to it
-sudo -u $SUDO_USER tmux start-server
-# create a new session but don't attach to it either
-sudo -u $SUDO_USER tmux new-session -d
-# install the plugins
-sudo -u $SUDO_USER /home/$SUDO_USER/.tmux/plugins/tpm/scripts/install_plugins.sh
-# killing the server is not required, I guess
-sudo -u $SUDO_USER tmux kill-server
+if [ ! -d "/home/$SUDO_USER/.tmux"]
+then
+    sudo -u $SUDO_USER git clone https://github.com/tmux-plugins/tpm /home/$SUDO_USER/.tmux/plugins/tpm
+    # start a server but don't attach to it
+    sudo -u $SUDO_USER tmux start-server
+    # create a new session but don't attach to it either
+    sudo -u $SUDO_USER tmux new-session -d
+    # install the plugins
+    sudo -u $SUDO_USER /home/$SUDO_USER/.tmux/plugins/tpm/scripts/install_plugins.sh
+    # killing the server is not required, I guess
+    sudo -u $SUDO_USER tmux kill-server
+fi
 
 # Soft link config files in home directory
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
